handler_class_note_split
