+++ assistant_bot/class_commands_handler_notes.py
@@ -1,9 +1,66 @@
 
 from .class_commands_handler import Commands_Handler
 from .class_notes_ext import Notes_Storage
+from functools import wraps
 
 class Commands_Handler_Notes(Commands_Handler):
     notes_storage: Notes_Storage
 
+
+    def backup_data_note(func):
+        @wraps(func)
+        def wrapper(self, *args, **kwargs):
+            result = func(self, *args, **kwargs)
+            if self and self.backup_data:
+                self.backup_data()
+            self.backup_data()
+            # if self._callback is not None:
+            #     self._callback("backup_data")
+            return result
+        return wrapper
+
+
+    @backup_data_note
+    @Commands_Handler.input_error
+    def handler_add_note(self, *args) -> str:
+        return self.notes_storage.add_note(*args)    
+
+
+    @backup_data_note
+    @Commands_Handler.input_error
+    def handler_change_notes(self, *args):
+        return self.notes_storage.change_note(*args)
+
+
+    @backup_data_note
+    @Commands_Handler.input_error
+    def handler_delete_notes(self, *args):
+        return self.notes_storage.delete_note(*args)
+
+
+    @backup_data_note
+    @Commands_Handler.input_error
+    def handler_clear_notes(self, *args):
+        return self.notes_storage.clear_notes(*args)
+
+
+    @Commands_Handler.input_error
+    def handler_search_notes(self, *args):
+        return self.notes_storage.search_notes(*args)
+
+
+    @Commands_Handler.input_error
+    def handler_search_notes(self, *args):
+        return self.notes_storage.search_notes(*args)
+
+
+    @Commands_Handler.input_error
+    def handler_sort_notes(self, *args):
+        return self.notes_storage.sort_notes(*args)
+
+
+    @Commands_Handler.input_error
+    def handler_show_notes(self, *args) -> str:
+        return self.notes_storage.show_notes()
